refactor(protocol): remove dead IND-543 shells and fix the tool-composition inversion
Stages 1-2 of the packages/protocol restructure. No public API change: all 441
exported symbols are byte-identical to 13.1.0 and src/index.ts is untouched.

Removed the IND-543 outer shells — src/public/, src/platform/,
src/runtime/foreground/, src/runtime/background/. All four were
declaration-only placeholders with zero inbound imports; runtime/background
and public contained nothing but a header comment. The only thing referencing
them was a spec asserting they existed.

Removed eight capability barrels (signals, communities, contacts, questions,
opportunity, negotiation, integrations, participant-agents). Each was
`export * from "./public/index.js"` with no importers — every real consumer
already resolved to public/ directly.

Fixed the tool-composition shim inversion: tool.registry.ts and tool.factory.ts
were implemented under runtime/foreground/composition/ and re-exported from
shared/agent/ through modules marked @deprecated, but all 25 importers used the
deprecated path. The implementations now live at the path everyone already
imported, and the indirection is gone. src/README.md already documented them at
shared/agent/ — the docs were right and the code had drifted.

Also removed src/shared/ui/ (no source, only a tests/tsconfig.json) and the
ambient-background / neutral-platform / public-compatibility capability
classifications, which existed solely to describe the deleted shells.

Replaced architecture/tests/runtime-shells.spec.ts with package-entry.spec.ts,
asserting the invariants that outlived the shells: src/index.ts is the sole
package.json export, and the composition root has exactly one implementation.

The protocol atlas cited runtime/background/index.ts — an empty comment-only
file — as evidence for background protocol behavior. Those citations now point
at the queue port and opportunity graph, which is where the behavior lives.

Verification:
- bun run build: clean
- public surface: 441 exported symbols byte-identical to baseline
- bun run architecture:check: host isolation OK, 24 named directions OK, 22 pass
- bun run check:protocol-atlas: artifact current (57 nodes, 80 edges)
- services/api tsc --noEmit: clean
- bun test src: 96 fail / 2 errors, identical to baseline on the same tree
  (pre-existing LLM-dependent failures); the 6-test delta is exactly the
  removed shell-existence assertions
