+++ docs/protocol-atlas/atlas-content.js
@@ -148,12 +148,12 @@
       title: "Commit and enqueue background work",
       summary: "The approved Signal is committed and deferred for background discovery processing.",
       conceptIds: ["signal", "opportunity"],
-      nodeIds: ["component.intent-graph-factory", "host-requirement.intent-graph-queue", "runtime-shell.background"],
+      nodeIds: ["component.intent-graph-factory", "host-requirement.intent-graph-queue"],
       invariantIds: ["participant-consent", "action-attribution", "host-boundary"],
-      sourcePaths: ["packages/protocol/src/signals/application/intent.graph.ts", "packages/protocol/src/shared/interfaces/queue.interface.ts", "packages/protocol/src/runtime/background/index.ts"],
+      sourcePaths: ["packages/protocol/src/signals/application/intent.graph.ts", "packages/protocol/src/shared/interfaces/queue.interface.ts"],
       notes: {
         protocol: "Current protocol behavior creates opportunities through background processing; synchronous examples are stale references, not an alternate normative path.",
-        implementation: "The package declares a queue port and background runtime surface; scheduling infrastructure is a host boundary.",
+        implementation: "The package declares a queue port; scheduling infrastructure is a host boundary.",
       },
     },
   ]);
@@ -164,12 +164,12 @@
       title: "Load the background trigger",
       summary: "Background processing starts from a committed Signal and its attributable participant context.",
       conceptIds: ["participant", "signal", "context"],
-      nodeIds: ["runtime-shell.background", "component.opportunity-graph-factory"],
+      nodeIds: ["component.opportunity-graph-factory"],
       invariantIds: ["action-attribution", "context-freshness"],
-      sourcePaths: ["packages/protocol/src/runtime/background/index.ts", "packages/protocol/src/opportunity/application/opportunity.graph.ts"],
+      sourcePaths: ["packages/protocol/src/opportunity/application/opportunity.graph.ts"],
       notes: {
         protocol: "A trigger identifies the Signal being served; it is not permission to search every community.",
-        implementation: "The background shell exposes ambient protocol composition and invokes the opportunity graph.",
+        implementation: "Host-scheduled background work invokes the opportunity graph; the package owns no scheduler.",
       },
     },
     {
@@ -365,9 +365,9 @@
       title: "Build the authorized tool registry",
       summary: "Only tool families admitted by the effective policy are made available for the request.",
       conceptIds: ["software-agent", "agent-permission", "effective-scope"],
-      nodeIds: ["component.tool-registry", "runtime-shell.foreground"],
+      nodeIds: ["component.tool-registry", "runtime-shell.composition"],
       invariantIds: ["scope-intersection", "action-attribution"],
-      sourcePaths: ["packages/protocol/src/runtime/foreground/composition/tool.registry.ts", "packages/protocol/src/runtime/foreground/index.ts"],
+      sourcePaths: ["packages/protocol/src/shared/agent/tool.registry.ts"],
       notes: {
         protocol: "Authorization filters capabilities before invocation rather than relying on an agent to avoid disallowed tools.",
         implementation: "The foreground composition registry assembles protocol-owned tool families.",
@@ -852,7 +852,7 @@
       { id: "gap-bounded-negotiation", kind: "discrepancy", title: "Bounded negotiation versus external-agent exception", summary: "Normative negotiation is bounded. The current package graph treats maxTurns=0 as uncapped when both participants have external agents; this is an implementation discrepancy, not a new rule.", sourcePaths: ["packages/protocol/src/negotiation/application/negotiation.graph.ts"] },
       { id: "gap-lifecycle-vocabulary", kind: "discrepancy", title: "Product and internal lifecycle vocabularies", summary: "Draft/Sent/Connected/Declined/Expired are product states, while the package uses internal lifecycle states. The atlas maps both and does not invent a merged lifecycle.", sourcePaths: ["packages/protocol/src/opportunity/application/opportunity.tools.ts"] },
       { id: "gap-community-network", kind: "discrepancy", title: "Community versus Network/index", summary: "Community is the normative protocol concept; Network is product language and network/index is current implementation vocabulary.", sourcePaths: ["packages/protocol/src/communities/application/network.graph.ts", "packages/protocol/src/communities/application/indexer.graph.ts"] },
-      { id: "gap-background-discovery", kind: "discrepancy", title: "Background discovery versus stale synchronous examples", summary: "Current opportunity creation is background-only. Synchronous examples are stale explanatory material and do not define an alternate protocol flow.", sourcePaths: ["packages/protocol/src/runtime/background/index.ts", "packages/protocol/src/opportunity/application/opportunity.graph.ts"] },
+      { id: "gap-background-discovery", kind: "discrepancy", title: "Background discovery versus stale synchronous examples", summary: "Current opportunity creation is background-only. Synchronous examples are stale explanatory material and do not define an alternate protocol flow.", sourcePaths: ["packages/protocol/src/opportunity/application/opportunity.graph.ts"] },
       { id: "gap-candidate-presentation", kind: "discrepancy", title: "Private Candidate versus surfaced Opportunity", summary: "Candidate identity and evaluation stay private; only a recipient-safe Opportunity presentation is surfaced. The two are not interchangeable records.", sourcePaths: ["packages/protocol/src/opportunity/application/opportunity.presenter.ts"] },
       { id: "reference-radar", kind: "reference-concept", title: "Radar", summary: "Product/reference-implementation retrieval concept; not a normative primitive." },
       { id: "reference-semantic-entropy", kind: "reference-concept", title: "Semantic entropy", summary: "Product/reference-implementation uncertainty technique; not a normative primitive." },
