+++ packages/protocol/src/participant-context/application/index.ts
@@ -89,7 +89,7 @@ export {
 } from "../../enrichment/enrichment.enricher.js";
 
 // ── Tool factories (foreground adapters) ──────────────────────────────────────
-// These are foreground adapter entry points consumed by runtime/foreground
+// These are foreground adapter entry points consumed by the tool composition root (shared/agent)
 // composition.  They accept compiled graphs and host deps through the tool-dep
 // ports and produce LangChain-compatible tool arrays.
 export { createEnrichmentTools } from "../../enrichment/enrichment.tools.js";
