+++ packages/protocol/src/questions/public/index.ts
@@ -9,7 +9,7 @@
  * ## Boundary
  *
  * References only questions/domain, questions/application, and questions/ports.
- * Never imports from runtime/foreground, host implementations, or other
+ * Never imports from the tool composition root (shared/agent), host implementations, or other
  * capability internals.
  *
  * ## Intentionally excluded from public surface
